requirements.txt
django
djangorestframework
cryptography
